+++ modules/request/index.js
@@ -1,9 +1,80 @@
-
+var request = require('request');
 var http = require('http');
 var https = require('https');
+var zlib = require('zlib');
 
 var gracenode = require('../../');
 var log = gracenode.log.create('request');
+
+/*
+options: {
+	headers: <object>,
+	encoding: <string>,
+	gzip: <boolean>
+}
+*/
+
+module.exports.GET = function (url, args, options, cb) {
+	send(url, 'get', args, options, cb);
+};
+
+module.exports.POST = function (url, args, options, cb) {
+	send(url, 'post', args, options, cb);
+};
+
+module.exports.PUT = function (url, args, options, cb) {
+	send(url, 'put', args, options, cb);
+};
+
+module.exports.DELETE = function (url, args, options, cb) {
+	send(url, 'del', args, options, cb);
+};
+
+function send(url, method, args, options, cb) {
+	if (!options) {
+		options = {};
+	}
+	var params = {
+		encoding: options.encoding || null,
+		url: url,
+		headers: options.headers || {}
+	};
+	if (method === 'get') {
+		params.form = args;
+	} else {
+		params.body = args;
+		params.json = true;
+	}
+
+	var sender = request[method] || null;
+	
+	if (!sender) {
+		return cb(new Error('invalid request method given: ' + method.toUpperCase()));
+	}
+
+	var gzip = options.gzip || false;
+
+	sender(params, function (error, res, body) {
+		if (error) {
+			return cb(error, body, res.statusCode);
+		}
+		if (!gzip) {
+			return cb(null, body, res.statusCode);
+		}
+		zlib.gunzip(body, function (err, unzipped) {
+			if (err) {
+				return cb(err);
+			}
+			unzipped = unzipped.toString();
+			try {
+				body = JSON.parse(unzipped);
+			} catch (e) {
+				return cb(new Error(unzipped), unzipped, res.statusCode);
+			}
+			cb(res.statusCode > 399 ? new Error(res.statusCode) : null, body, res.statusCode);
+		});
+	});
+}
 
 /**
  * @param {object} { protocol: http/https, hots: string, path: string. port: number, method: string, data: object }